requirements.txt
# Frozen from venv (pip freeze). Regenerate with: pip freeze > requirements.txt

a2a-sdk==0.3.24
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
backoff==2.2.1
beeai-framework==0.1.76
brotli==1.2.0
cachetools==7.0.1
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
chevron==0.14.0
choreographer==1.2.1
click==8.3.1
constantly==23.10.4
cryptography==46.0.5
cssselect2==0.9.0
daphne==4.2.1
ddgs==9.11.1
Deprecated==1.3.1
distro==1.9.0
Django==5.2.11
django-prometheus==2.4.1
faiss-cpu==1.13.2
fastapi==0.129.2
fastuuid==0.14.0
filelock==3.24.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
google-api-core==2.30.0
google-auth==2.48.0
googleapis-common-protos==1.72.0
grpcio==1.78.1
grpcio-health-checking==1.78.1
grpcio-reflection==1.78.1
grpcio-tools==1.78.1
h11==0.16.0
hf-xet==1.2.0
hiredis==3.3.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.4.1
hyperlink==21.0.0
idna==3.11
importlib_metadata==8.7.1
Incremental==24.11.0
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
json_repair==0.52.5
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kaleido==1.2.0
langchain-core==1.2.13
langchain-text-splitters==1.1.0
langfuse==3.14.1
langsmith==0.7.3
litellm==1.81.11
logistro==2.0.1
lxml==6.0.2
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
narwhals==2.17.0
networkx==3.6.1
numpy==2.4.2
openai==2.21.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
packaging==25.0
pandas>=2.0.0
pillow==12.1.1
plotly==6.6.0
pluggy==1.6.0
primp==1.1.2
prometheus_client==0.24.1
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydyf==0.12.1
Pygments==2.19.2
pyOpenSSL==25.3.0
pypdf==6.7.0
pyphen==0.17.2
pytest==9.0.2
pytest-timeout==2.4.0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.1.1
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
ruff==0.15.4
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
service-identity==24.2.0
setuptools==82.0.0
shellingham==1.5.4
simplejson==3.20.2
sniffio==1.3.1
sqlparse==0.5.5
sse-starlette==3.2.0
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.5.1
tinyhtml5==2.1.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.2.0
Twisted==25.5.0
txaio==25.12.2
typer==0.23.1
typer-slim==0.23.1
types-grpcio==1.0.0.20251009
types-grpcio-health-checking==1.0.0.20250506
types-grpcio-reflection==1.0.0.20250506
types-protobuf==6.32.1.20260221
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.37.0
uvloop==0.22.1
watchfiles==1.1.1
weasyprint==68.1
webencodings==0.5.1
websockets==16.0
whitenoise==6.11.0
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zope.interface==8.2.0
zopfli==0.4.1
zstandard==0.25.0
